LPS-186836 Extract service from WikiPageIndexer to WikiPageRelatedEntryIndexer and extends BaseRelatedEntryIndexer in it
